add x axis and resolve type error
